% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.timepoints.R
\name{parse.timepoints}
\alias{parse.timepoints}
\title{Apply a closure function generated by create.timepoint.parser to leaf names.}
\usage{
parse.timepoints(seq_names = NULL, timepoints_parser = NULL, uniquify = F,
  refseq_name = NULL, name_list = F)
}
\arguments{
\item{seq_names}{A vector or list of type character, for parsing.}

\item{timepoints_parser}{A closure function.}

\item{uniquify}{Option to return a normalized (non-redundant) list of parsed entries.}

\item{refseq_name}{If this sequence name is in the alignment, its name will not be parsed, but the corresponding entry will be assigned NULL instead.}

\item{name_list}{Should the resulting list be named after the input sequences?}
}
\value{
A (possibly named) list of fields parsed from the input sequence names.
}
\description{
Apply a closure function generated by create.timepoint.parser to leaf names.
}

