% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixgramr.R
\docType{package}
\name{pixgramr}
\alias{pixgramr}
\alias{pixgramr-package}
\title{pixgramr: Display a pixel plot and phylogram of aligned sequences.}
\description{
The two principal methods are aconstructor, which usies the S3 object model and  a plotting function.  For details about how to use each, run \code{help(pixgramr::pixgram)} and \code{help(pixgramr::plot.pixgram)}.
}
\details{
For examples run \code{vignette("pixgramr")}.
}
\seealso{
\code{\link{pixgram}} \code{\link{plot}}
}

