% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixmap.colors.R
\docType{data}
\name{pixmap_colors}
\alias{pixmap_colors}
\title{Color Look-up-Table}
\format{An object of class \code{data.frame} with 36 rows and 4 columns.}
\usage{
pixmap_colors
}
\description{
Used internally depict aas or nts by category.
To customize the color scheme, edit 'color-lut.csv' and either (a)
uncomment the 'overwrite' argument then rebuild, reinstall, and restart;
or (b) read the modified table into memory and map it to the color_lut slot
in the pixelgram object when instantiating or plotting.
}
\keyword{datasets}

