% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pixelgram}
\alias{plot.pixelgram}
\title{Renders the pixelgram object}
\usage{
\method{plot}{pixelgram}(x, xform_type = NULL, xform_master = NULL,
  leaf_colors = NULL, tip_labels = NULL, legend_attributes = NULL,
  edge_widths = NULL, edge_colors = NULL, scale_bar = NULL,
  pheno_matrix = NULL, pheno_colnames = NULL, pheno_letters = NULL,
  show_top_axis = F, show_tip_label = F, selected_rows = NULL,
  selected_row_colors = NULL, no_margin = F, plot_margin_points = F,
  vbars = NULL, raster_width = NULL, raster_margin = NULL,
  color_lut_type = NULL, notes = NULL, annotate_env = F,
  show_tree = NULL, main = NULL, sub = NULL, ...)
}
\arguments{
\item{x}{pixelgram object to be plotted}

\item{xform_type}{see \code{help(pixelgram::pixelgram())}.}

\item{xform_master}{see \code{help(pixelgram::pixelgram())}.}

\item{leaf_colors}{vector of tree leaf colors; can be colors or integers}

\item{tip_labels}{List for adding tree tip labels; see details.}

\item{legend_attributes}{Named list of entries that go into the legend}

\item{edge_widths}{tree edge widths}

\item{edge_colors}{tree edge colors}

\item{scale_bar}{Named list of entries used to draw the scale bar}

\item{pheno_matrix}{Heatmap of RGB values to draw in raster_margin.}

\item{pheno_colnames}{A vector of names to label pheno_matrix.}

\item{pheno_letters}{A vector of letters to label pheno_matrix.}

\item{show_top_axis}{Draw axis along top of plot?}

\item{show_tip_label}{Show taxon labels on tree tips?}

\item{selected_rows}{Vector of selected taxa.}

\item{selected_row_colors}{Colors to depict vector of selected taxa.}

\item{no_margin}{No margin?  This is passed to \code{ape::plot.phylo()}.}

\item{plot_margin_points}{Plot margin points?}

\item{vbars}{vertical bars?}

\item{raster_width}{Proportionate width of the pixel plot, e.g. 1 indicates equal width.}

\item{raster_margin}{Offset in raster portion of the layout, e.g. for heatmap.}

\item{color_lut_type}{Optional string for non-default color lookup table, currently only implemented for amino acids as 'charge' and 'taylor'.}

\item{notes}{Is a named list used to draw landmarks for the region sequenced, with elements "Lhs", "Rhs", "clr" and "txt".  Using numbering from the reference sequence lookup-table, rectangles are drawn bounded by Lhs and Rhs for each entry and filled with the color specified by clr (using alpha transparency gives better results).  If present, labels with values in "txt" are plotted along the margin.  If annotate_env is true, these values are populated internally for the HIV-1 env.}

\item{annotate_env}{If true, draw Env landmarks.}

\item{show_tree}{Show the tree? If not, only show the pixel plot.}

\item{main}{Plot title, e.g. coded subject id.  If given, it is plotted at the top of the pixelgram output.}

\item{sub}{Plot subtitle, i.e. region sequenced.  If specified, it appears with "site" below the pixel plot.}

\item{...}{dots, passed to ape::plot.phylo()}
}
\description{
This layout-compatible method validates the pixelgram object, establishes plot coordinates, and draws the pixel and phylogram plots.
}
\details{
The x-axis of the plot coordinate system is positive for the tree-plotting region, and negative for the pixel-plotting region.  Setting \code{raster_width} rescales the minimum x-axis value by this factor.  The default is 1, so the pixel and raster plots are equal width.  The \code{raster_margin} parameter rescales the right (least negative) edge of the pixel plot, to leave a gap between pixel and tree, e.g. to include a heatmap of phyenotypic assay results.

Tree tip labels can be added by passing a list called \code{tip_labels}, with elements named "pch", "bgs", and "col".  See vignette.  The tree can be further decorated by passing \code{leaf_colors}, \code{edge_widths}, and \code{edge_colors}.  See the last vignette example.

If defined, data in the \code{pheno_matrix} are plotted in the margin between tree and pixel plots.  Entry values should be positive (non-zero) integers, which are interpreted as indices to \code{palette()}, color names, or color values.  Matrix rows must have names that match tree tip-label (i.e. \code{pg$tre$tip.label}  strings, or they will not be plotted.  This is to facilitate their proper ordering, which follows that of the leaves on the tree.  All columns are plotted, and in the order given.  However, column widths are scaled to fit together into the marginal area provided within the pixel-plot area (x-values < 0) by \code{raster_margin}.

The x- and y-axis lim arguments are provided to facilitate plotting multiple trees on the same scale in different panels.

The option \code{color_lut_type} is used to specify alternative columns from the defined color scheme (see \code{help(pixmap_colors)}).  For example, 'taylor' in the vignette uses the colors described in Wittgensteinean fashion by W. Taylor, Protein Engineering, Vol 10 , 743-746 (1997).

Entries in the \code{scale_bar} list argument should be named \code{'pos'} for plot-position coordinates, \code{'lwd'}, \code{'len'} for bar length, in terms of the plot coordinates, \code{'cex'}, and \code{'text'}.

Entries in the \code{legend_attributes} list argument should be named \code{'location'} for legend positioning as defined in \code{?legend}, \code{'pch'}, \code{'cex'}, \code{'pt.cex'}, \code{'lwd'}, \code{'bg'}, \code{'bty'}, \code{'box.col'}, \code{'col'} for colors, \code{'title'}, and \code{'text'}.
}
\examples{
\dontrun{
  plot( pg <- pixelgram::pixelgram(nts=pixelgram::hiv.ref$nts) )
}
}

