% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pixgram}
\alias{plot.pixgram}
\title{Renders the pixgram object}
\usage{
\method{plot}{pixgram}(x, xform_type = NULL, xform_master = NULL,
  leaf_colors = NULL, tip_labels = NULL, legend_attributes = NULL,
  edge_widths = NULL, edge_colors = NULL, scale_bar = NULL,
  pheno_matrix = NULL, show_top_axis = F, show_tip_label = F,
  selected_rows = NULL, no_margin = F, plot_margin_points = F,
  vbars = NULL, raster_width = NULL, raster_margin = NULL,
  color_lut_type = NULL, annotate_env = F, ...)
}
\arguments{
\item{x}{pixgram object to be plotted}

\item{xform_type}{see \code{help(pixgramr::pixgram())}.}

\item{xform_master}{see \code{help(pixgramr::pixgram())}.}

\item{leaf_colors}{vector of tree leaf colors; can be colors or integers}

\item{tip_labels}{vector of tree tip labels}

\item{legend_attributes}{Named list of entries that go into the legend}

\item{edge_widths}{tree edge widths}

\item{edge_colors}{tree edge colors}

\item{scale_bar}{Named list of entries used to draw the scale bar}

\item{pheno_matrix}{heatmap to draw in raster_margin}

\item{show_top_axis}{Draw axis along top of plot?}

\item{show_tip_label}{Show taxon labels on tree tips?}

\item{selected_rows}{Vector of selected taxa}

\item{no_margin}{No margin?  This is passed to \code{ape::plot.phylo()}.}

\item{plot_margin_points}{Plot margin points?}

\item{vbars}{vertical bars?}

\item{raster_width}{Proportionate width of the pixel plot, e.g. 1 indicates equal width.}

\item{raster_margin}{Offset in raster portion of the layout, e.g. for heatmap.}

\item{color_lut_type}{Optional string for non-default color lookup table, currently only implemented for amino acids as 'charge' and 'taylor'.}

\item{annotate_env}{If true, draw Env landmarks.}
}
\description{
This layout-compatible method validates the pixgram object, establishes plot coordinates, and draws the pixel and phylogram plots.
}
\details{
The x-axis of the plot coordinate system is positive for the tree-plotting region, and negative for the pixel-plotting region.  Setting \code{raster_width} rescales the minimum x-axis value by this factor.  The default is 1, so the pixel and raster plots are equal width.  The \code{raster_margin} parameter rescales the right (least negative) edge of the pixel plot, to leave a gap between pixel and tree, e.g. to include a heatmap of phyenotypic assay results.

The tree can be decorated by passing \code{leaf_colors}, \code{tip_labels}, \code{edge_widths}, and \code{edge_colors}.  See the last vignette example.

If defined, data in the \code{pheno_matrix} are plotted in the margin between tree and pixel plots.  Entry values should be positive (non-zero) integers, which are interpreted as indices to \code{palette()}, color names, or color values.  Matrix rows must have names that match tree tip-label (i.e. \code{pg$tre$tip.label}  strings, or they will not be plotted.  This is to facilitate their proper ordering, which follows that of the leaves on the tree.  All columns are plotted, and in the order given.  However, column widths are scaled to fit together into the marginal area provided within the pixel-plot area (x-values < 0) by \code{raster_margin}.

The x- and y-axis lim arguments are provided to facilitate plotting multiple trees on the same scale in different panels.

The option \code{color_lut_type} is used to specify alternative columns from the defined color scheme (see \code{help(pixmap_colors)}).  For example, 'taylor' in the vignette uses the colors described in Wittgensteinean fashion by W. Taylor, Protein Engineering, Vol 10 , 743-746 (1997).

Entries in the \code{scale_bar} list argument should be named \code{'pos'} for plot-position coordinates, \code{'lwd'}, \code{'len'} for bar length, in terms of the plot coordinates, \code{'cex'}, and \code{'text'}.

Entries in the \code{legend_attributes} list argument should be named \code{'location'} for legend positioning as defined in \code{?legend}, \code{'pchs'}, \code{'cex'}, \code{'pt.cex'}, \code{'lwds'}, \code{'bg'}, \code{'bty'}, \code{'box.col'}, \code{'cols'} for colors, \code{'title'}, and \code{'text'}.
}
\examples{
\dontrun{
  plot( pg <- pixgramr::pixgram(nts=pixgramr::hiv.ref$nts) )
}
}

