% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixmap.colors.R
\docType{data}
\name{pixmap_colors}
\alias{pixmap_colors}
\title{Color Look-up-Table}
\format{\preformatted{'data.frame':	36 obs. of  4 variables:
 $ nt    : chr  "red" "blue" "orchid" "orchid" ...
 $ aa    : chr  "#FFCC33" "#CC3333" "#3333CC" "#3333CC" ...
 $ charge: chr  "grey42" "grey42" "red" "red" ...
 $ taylor: chr  "#CCFF00" "#FFFF00" "#FF0000" "#FF0066" ...
}}
\usage{
pixmap_colors
}
\description{
Used internally depict aas or nts by category.
To customize the color scheme, edit 'color-lut.csv' and either (a)
uncomment the 'overwrite' argument then rebuild, reinstall, and restart;
or (b) read the modified table into memory and map it to the color_lut slot
in the pixgram object when instantiating or plotting.
}
\keyword{datasets}

