% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixgram.R
\name{set.aln.file}
\alias{set.aln.file}
\title{Sets the alignment file, whether nt or aa.  Also sets the refseq.}
\usage{
set.aln.file(P, f, file_type = "aa", alignment_format = "fasta")
}
\arguments{
\item{pixgram}{object}

\item{alignment}{file name}

\item{alignment}{file format}

\item{alignment}{file type}
}
\value{
updated pixgram object
}
\description{
Sets the alignment file, whether nt or aa.  Also sets the refseq.
}

