% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixelgram.default.R
\name{set.aln.file}
\alias{set.aln.file}
\title{Sets the alignment file, whether nt or aa.  Also sets the refseq.}
\usage{
set.aln.file(P, f, file_type = "aa", alignment_format = "fasta")
}
\arguments{
\item{P}{pixelgram object}

\item{f}{alignment file name}

\item{file_type}{Specifies alignment file type.  Must be "aa", "nt", or "codon".}

\item{alignment_format}{Alignment file format, one "fasta", "clustal", "phylip", "msf", or "mase"}
}
\value{
updated pixelgram object
}
\description{
Sets the alignment file, whether nt or aa.  Also sets the refseq.
}

