% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixgram.R
\name{set.tre.file}
\alias{set.tre.file}
\title{Assign a Newick-formatted tree file to a pixgram object.}
\usage{
set.tre.file(P, f)
}
\value{
Updated pixgram object
}
\description{
Assign a Newick-formatted tree file to a pixgram object.
}

