% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pixelgram.default.R
\name{set.tre.file}
\alias{set.tre.file}
\title{Assign a Newick-formatted tree file to a pixelgram object.}
\usage{
set.tre.file(x, f)
}
\arguments{
\item{x}{Pixelgram object}

\item{f}{Name of file containing a Newick tree.}
}
\value{
Updated pixelgram object
}
\description{
Assign a Newick-formatted tree file to a pixelgram object.
}

